3-4주차 수정
